Fix test when removing an open file
When testing the use of the `append=true` option to `write_data`,
we previously call `rm` on the path returned by `mktemp`
inside the function passed to `mktemp`.  While macOS doesn't
mind this, most systems do not.  Rewrite the test so that we
no longer delete an open file then rewrite to it to avoid
this problem.