Fix failing tests for overwriting new files
We test that calling `LibMseed.write_data(file, ...; overwrite=true)`
works as expected when `file` does not already exist.  However
on Ubuntu 20.04 and Julia v1.7.2, this was causing a problem which
varied depending on whether the platform was 32- or 64-bit, perhaps
related to the fact we quickly write to a file, then `rm(file)`,
then call `write_data` again with `append=true`.

Try to avoid this problem by appending to a non-existent file
which has not been recently delted.