Properly GC root C-allocated structs when reading
It was possible for Julia to free a reference to a libmseed
`MseedTraceList` in `read_file` and maybe `read_buffer` before
calling `ms3_readtracelist`, manifesting itself as incorrect
file position identifiers in the C structs, which then lead to
an error when reading files using those incorrect pointers.
Fix the issue by `GC.@preserve`ing the Julia pointers to those
C structs when reading.