Return documented things in channel_code_parts and don't throw errors
Previously, `channel_code_parts`'s returned elements of a channel code
as `nothing` if they were empty in a miniSEED code string, or if the
format was not as expected.  However, the docstring actually stated they
would be 'empty', meaning empty strings.  Update the behaviour to match
that documented.

At the same time, also stop erroring when encountering a channel code
which for any reason does not fit the expected formats.  This is much
preferable because otherwise it is impossible to get the string out
of LibMseed.