Add `headers_only` option to `read_file` and `read_buffer`
When `headers_only` is `true`, then we do not pass the flags to
the underlying libmseed library to unpack the data, meaning the
trace data is not read into memory.  This can be useful when only
trying to find out trace information and not the contents, for
example if determining the start and end time or sampling rate of
traces.

When only headers are read, the data element type of the trace is
set to `Missing` and the data vector is empty.