Fix tests using `time_tolerance` on unsupported platforms
The `time_tolerance` option to `read_file` and `read_buffer` is not
supported on PowerPC and ARM platforms, but we used the option
unconditionally on all platforms in tests.  Fix this to instead test
for the correct throwing of an error on ARM and PowerPC and skip the
`time_tolerance` tests.