Enforce NanosecondDateTimes to be in the right range
`NanosecondDateTime`s in LibMseed.jl might have any value that a
`DateTime` can have, but since libmseed's dates are in ns since the
epoch, `NanosecondDateTime`s shouldn't be able to represent that.  Make
sure that they cannot, and avoid wraparound when converting a
`NanosecondDateTime` to the libsmeed `nstime_t` type, which is just an
`Int64`.