Implement selections when reading
libmseed allows for 'selections', or reading only a subset of data
matching a channel ID globbing pattern and a start and end date.
Implement both of these via the `channels`, `startdate` and `enddate`
keyword arguments to `read_buffer` and `read_file`.

libmseed returns a 'not SEED data' error when reading a valid file but
where no traces match the selection criteria.  We choose to return an
empty `LibMseed.MseedTraceList` in this case (where some selection
criteria are set), but it is possible therefore to read an invalid file
with selection criteria and not throw an error.