Fix incorrect error handling in read_buffer
It was previously assumed that `mstl3_readbuffer_selection` would
behave the same as `ms3_readtracelist_selection` and return an
error if a buffer was valid but did not contain any traces which
match the selection criteria.  In fact, the former function
does not return an error, but rather a trace list with no traces
(which makes more sense).  Fix the logic by removing a branch
which would only ever be taken erroneously.