Clarify station ID form for `channels` in `read_{file,buffer}`
Make it clear that the `channels` string used to match channels in
`read_buffer` and `read_file` is matched against the full source
ID in FDSN form, and add a link to the documentation for this.

Also make it clear that it is expected that an error is printed to
screen in `read_file` when no data match the selections.